python/gui/requirements.txt
pyinstaller 
pyside6
python-ffmpeg